Create fixtures prior to starting watcher
This test worked before by exploiting an inefficiency in chokidar (an
unnecessary fs.realpath call), which was removed in the prior commit.
This could also be made to work by adding a delay between creating
watcher, and writing the fixtures, but dealing with watching previously
non-existent paths is not what this test is about.