check the real path for a watched parent
This change creates fs instances for symlinks that link out to a file not in an already watched directory which is the parent of the symlink. Specifically this fixes watching files that have been npm link'd.

There might be a more efficient answer, that creates an fs instance for the symlinked directory, but as far as i could tell the only reason, my parent directory is currently being watched is because karma-browserify specifically includes the package.json as a watched file. That is, it doesn't seem to be something chokidar is doing intentionally.