Adjust testing watch mode logic
Skip testing fs.watch mode on OS X & node 0.10 entirely. It’s not the
default or even the fallback mode, and it is too flaky to get all the
tests passing consistently on it in CI environments. I’ve wasted too
much of my life wrestling with this already.

Leave specific osXFsWatch010 conditions in place for a while at least
just in case this decision gets reversed.