feat: rework anymatch and fix ignored paths
This reworks the anymatch function to be a little less generic since we
now know exactly what we want to call it with. This means we have much
clearer/stronger types.

On top of that, support for ignored paths has been implemented again (it
was removed with a `TODO` until now, in v4).

Keep in mind the way we store ignored paths may change in future so we
can more efficiently access and remove them without the need for full
iterations through the set.