feat: rework ignore matcher add/remove & add types
This basically does two things:

1. Adds `WatchHelper` to some of the methods for a stronger type
2. Reworks how we add/remove ignore paths ("matchers")

In the latter, thanks to a previous change where we can have "matcher
objects", we need to iterate through existing ignore paths to make sure
we don't duplicate them.

Similarly, when removing a "path", we need to find any matcher objects
which use the same path and remove those too.

Basically introducing `addIgnorePath` and `removeIgnorePath` instead of
mutating the set directly.