fix: revert default interval
Trying to fix some tests, the `interval` used to be `undefined` by
default. That mean we would later compare `n > undefined` (which would
be false). Changing it to `0` means `n > 0` would now be true.

This doesn't actually fix the tests most likely but may as well revert
to the same behavior for now.