fix: mac OS inodes change just like linux
When a file changes in linux, it gets a new inode which means our
`fs.watch` will still be watching the old inode, not the new one.

To get around this, we already remove the watcher and add a new one for
the new inode.

However, we don't apply this logic to MacOS, which also has the same
behaviour (creates a new inode when a file changes in some cases).

This change just makes it so we run the same logic for linux _and_ mac.