fix: add fsevents types
A bit of chicken-and-egg situation means we can't pull the official
fsevents types.

Basically, fsevents will only install on certain platforms (macOS?).

Of course, if we're not on a supported platform, that also means we
won't have access to the shipped typescript types.

To solve this, we have to keep a copy of the types ourselves.