test: fix symlinks on windows during testing
On windows, a symlink can be created as a `file`, a `dir`, or a
`junction`. Meanwhile, on unix-like systems, this type doesn't seem to
matter much (or is ignored).

By default, the type will be `file` on windows.

Due to this, all directories we symlink in tests are actually created as
"files" which leads to all sorts of permissions errors.

This change basically makes it explicit that the symlinked directories
are in fact directories (`dir`), thus making tests pass.