test: wait for each write to trigger an event
This test is incredibly flaky across different platforms, node versions,
etc when we try write all the files at once. Instead, we now write a
file at a time and wait for the spy to have been called before moving
on.

Not quite as accurate as real world but seems to be the only way to make
this stable.